+++ database/migrations/2020_05_01_122357_create_withdrawals_table.php
@@ -15,15 +15,15 @@ class CreateWithdrawalsTable extends Migration
     {
         Schema::create('withdrawals', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('member_id');
+            $table->bigInteger('member_id');
             $table->decimal('amount', 8, 2)->default(0);
             $table->decimal('tds_percentage', 8, 2)->default(0);
             $table->decimal('tds_amount', 8, 2)->default(0);
             $table->decimal('net_amount', 8, 2)->default(0);
-            $table->integer('withdrawal_request_id');            
+            $table->bigInteger('withdrawal_request_id');            
             $table->timestamp('payment_made_at')->nullable();
             $table->string('payment_status',32)->nullable();
-            $table->integer('transaction_by')->nullable();
+            $table->bigInteger('transaction_by')->nullable();
             $table->string('note',2048)->nullable();
             $table->timestamps();
             $table->softDeletes();
